spi: dt-bindings: ti,qspi: convert to dtschema
Convert txt binding of TI's qspi controller (found on their omap SoCs) to
dtschema to allow for validation.

It is however to be noted that it is not a one-to-one conversion, in the
sense that the original txt binding needed to be updated, but these
changes are included in the dtschema and are mentioned below.

The changes, w.r.t. the original txt binding, are:

- Introduce "clocks" and "clock-names" which was never mentioned.
- Reflect that "ti,hwmods" is deprecated and is not a "required"
  property anymore.
- Introduce "num-cs" which allows for setting the number of chip
  selects.

Signed-off-by: Kousik Sanagavarapu <five231003@gmail.com>